renamed main_utils to shared_utils since these are utils used in many files across the project
