+++ src/bot/commands.py
@@ -2,9 +2,9 @@ import discord
 from discord.ext import commands
 
 from ..buttons.base_view import create_base_view
-from ..main_utils.constants import COMMUNITY_SUPPORT_CHANNEL_ID, TROUBLESHOOTING_CHANNEL_ID
-from ..main_utils.enums import ThreadState
-from src.main_utils.shared_state import set_thread_state
+from ..shared_utils.constants import COMMUNITY_SUPPORT_CHANNEL_ID, TROUBLESHOOTING_CHANNEL_ID
+from ..shared_utils.enums import ThreadState
+from src.shared_utils.shared_state import set_thread_state
 import logging
