+++ src/shared_utils/constants.py
