+++ src/shared_utils/enums.py
