+++ src/shared_utils/online_resources.py
