+++ src/shared_utils/shared_state.py
@@ -1,4 +1,4 @@
-from src.main_utils.enums import ThreadState
+from src.shared_utils.enums import ThreadState
 
 thread_states = {}
